Added simple version number.
Now invokes the smart_bin function and seems to generate reasonable
messages. The types and widths of binary data are still a bit ad hoc
in their presentation, and could be simplified, but the result can be
deciphered.

Relies on a couple of assumptions: default type and default width.
As I recall these are unsigned-int and 8 bits?

Added more test cases to ex3.erl.